requirements.txt
tensorflow==2.16.2
tensorflow-io-gcs-filesystem==0.37.1
pandas
pillow
uvicorn
fastapi
google-cloud-storage==3.1.1
python-multipart
